backend/requirements-utf8.txt
absl-py==2.3.1
aiosmtplib==5.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asttokens==3.0.1
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
comm==0.2.3
cryptography==46.0.4
debugpy==1.8.19
decorator==5.2.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.128.0
Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.7.0
jupyter_core==5.9.1
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mediapipe==0.10.31
nest-asyncio==1.6.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.5
passlib==1.7.4
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.0
pure_eval==0.2.3
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
pyzmq==27.1.0
rsa==4.9.1
six==1.17.0
sounddevice==0.5.3
SQLAlchemy==2.0.45
stack-data==0.6.3
starlette==0.50.0
tornado==6.5.4
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.40.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==16.0
Werkzeug==3.1.4
